using fast static callbacks with SpanAction
