+++ src/EfficientDynamoDb/Internal/Extensions/Utf8JsonWriterExtensions.cs
@@ -54,29 +54,7 @@ namespace EfficientDynamoDb.Internal.Extensions
                 return;
             }
 
-            var tableNameContext = new TableNameFormatterContext(tableName);
-            var length = tableNameFormatter.CalculateLength(ref tableNameContext);
-
-            char[]? pooledArray = null;
-            var arr = length < NoAllocStringBuilder.MaxStackAllocSize
-                ? stackalloc char[length]
-                : pooledArray = ArrayPool<char>.Shared.Rent(length);
-
-            try
-            {
-                if( !tableNameFormatter.TryFormat(arr, ref tableNameContext, out length) ) {
-                    throw new DdbException($"Couldn't format table name '{tableName}' using the provided formatter");
-                }
-                writer.WriteString(tableNameKey, arr);
-            }
-            finally
-            {
-                if (pooledArray != null)
-                {
-                    pooledArray.AsSpan(0, length).Clear();
-                    ArrayPool<char>.Shared.Return(pooledArray);
-                }
-            }
+            tableNameFormatter.WriteTableName(tableName, writer, (arr, w) => w.WriteString(tableNameKey, arr));
         }
 
         /// <summary>
