+++ src/EfficientDynamoDb/Internal/Operations/Shared/BatchItemHttpContent.cs
@@ -1,9 +1,5 @@
-using System;
-using System.Buffers;
 using System.Text.Json;
 using EfficientDynamoDb.Configs;
-using EfficientDynamoDb.Exceptions;
-using EfficientDynamoDb.Internal.Core;
 
 namespace EfficientDynamoDb.Internal.Operations.Shared
 {
@@ -22,31 +18,7 @@ namespace EfficientDynamoDb.Internal.Operations.Shared
             }
 
 
-
-            var tableNameContext = new TableNameFormatterContext(tableName);
-            var length = tableNameFormatter.CalculateLength(ref tableNameContext);
-
-            char[]? pooledArray = null;
-            var arr = length < NoAllocStringBuilder.MaxStackAllocSize
-                ? stackalloc char[length]
-                : pooledArray = ArrayPool<char>.Shared.Rent(length);
-
-            try
-            {
-                if( !tableNameFormatter.TryFormat(arr, ref tableNameContext, out length) ) {
-                    throw new DdbException($"Couldn't format table name '{tableName}' using the provided formatter");
-                }
-
-                writer.WritePropertyName(arr[..length]);
-            }
-            finally
-            {
-                if (pooledArray != null)
-                {
-                    pooledArray.AsSpan(0, length).Clear();
-                    ArrayPool<char>.Shared.Return(pooledArray);
-                }
-            }
+            tableNameFormatter.WriteTableName(tableName, writer, (arr, w) => w.WritePropertyName(arr));
         }
     }
 }
